requirements.txt
pandas
requests
beautifulsoup4
pyarrow
google-cloud-storage
google-cloud-bigquery
pandas-gbq
python-dotenv
google-generativeai